+++ CHANGES.rst
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`_,
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
+Added
+*****
+
+- Configuration entries can be loaded from files if the entry key has a *_FILE* suffix
+  and the entry value is the path to the file. :issue:`134` :pr:`134`
+
 Removed
 *******